app/BE/requirements.txt
# backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0
apscheduler==3.10.1